글 작성, 글 수정 권한 체크, 글 삭제 권한 체크 시에 세션의 정보 사용

+++ src/main/java/com/yamto/example/jspCommunity/controller/UsrArticleController.java
@@ -78,7 +78,7 @@ public class UsrArticleController {
 		}
 		
 		int boardId = Integer.parseInt(req.getParameter("boardId"));
-		int memberId = Integer.parseInt(req.getParameter("memberId"));
+		int memberId = (int) session.getAttribute("loginedMemberId");
 		String title = req.getParameter("title");
 		String body = req.getParameter("body");
 		
@@ -116,6 +116,14 @@ public class UsrArticleController {
 			return "common/redirect";
 		}
 		
+		int loginedMemberId = (int) session.getAttribute("loginedMemberId");
+		
+		if(article.getMemberId() != loginedMemberId) {
+			req.setAttribute("alertMsg", id + "번 게시물에 대한 권한이 없습니다.");
+			req.setAttribute("historyBack", true);
+			return "common/redirect";
+		}
+		
 		articleService.delete(id);
 		
 		int boardId = article.getBoardId();
@@ -141,6 +149,14 @@ public class UsrArticleController {
 		
 		if(article == null) {
 			req.setAttribute("alertMsg", id + "번 게시물은 존재하지 않습니다.");
+			req.setAttribute("historyBack", true);
+			return "common/redirect";
+		}
+		
+		int loginedMemberId = (int) session.getAttribute("loginedMemberId");
+		
+		if(article.getMemberId() != loginedMemberId) {
+			req.setAttribute("alertMsg", id + "번 게시물에 대한 권한이 없습니다.");
 			req.setAttribute("historyBack", true);
 			return "common/redirect";
 		}
@@ -172,9 +188,9 @@ public class UsrArticleController {
 			return "common/redirect";
 		}
 		
-		int memberId = Integer.parseInt(req.getParameter("memberId"));
+		int loginedMemberId = (int) session.getAttribute("loginedMemberId");
 		
-		if(article.getMemberId() != memberId) {
+		if(article.getMemberId() != loginedMemberId) {
 			req.setAttribute("alertMsg", id + "번 게시물에 대한 권한이 없습니다.");
 			req.setAttribute("historyBack", true);
 			return "common/redirect";
